application_budget: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data

ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.

Usage

application_budget

Format

A data frame with 78 rows and 16 variables:

project_id

Identifier for each project within the ETH Open Research Data dataset. (Examples: CONTRIBUTE/PYCSOU-FAIR, EXPLORE/ASTROORDAS, CONTRIBUTE/CATCHARACTPROVI)

total_budget_personnel_senior_staff

Total budget allocated for senior staff personnel in Swiss Francs. (Range: 0 to 786000)

total_budget_personnel_postdocs

Total budget allocated for postdoctoral personnel in Swiss Francs. (Range: 0 to 309000)

total_budget_personnel_students

Total budget allocated for student personnel in Swiss Francs. (Range: 0 to 43535)

total_budget_personnel_other

Total budget allocated for other personnel in Swiss Francs. (Range: 0 to 144000)

total_budget_personnel_total_direct

Total direct budget allocated for personnel in Swiss Francs. (Range: 0 to 786000)

total_budget_travel

Total budget allocated for travel expenses in Swiss Francs. (Range: 0 to 13970)

total_budget_equipment

Total budget allocated for equipment expenses in Swiss Francs. (Range: 0 to 168000)

total_budget_other_publication_fees

Total budget allocated for other publication fees in Swiss Francs. (Examples: -1, 0, 9000)

total_budget_other_conferences_workshops

Total budget allocated for other conferences and workshops in Swiss Francs. (Range: 0 to 12500)

total_budget_other_other

Total budget allocated for other expenses in Swiss Francs. (Examples: 3000, 0, NA)

total_budget_other_total_direct

Total direct budget allocated for other expenses in Swiss Francs. (Range: 0 to 176000)

total_budget_total_direct

Total direct budget allocated for all expenses in Swiss Francs. (Range: 0 to 8e+05)

total_budget_subcontracting

Total budget allocated for subcontracting expenses in Swiss Francs. (Examples: NA, 30000, 25000)

total_budget_total_costs

Total budget allocated for all costs in Swiss Francs. (Range: 15000 to 1500000)

phase

Current phase of the project represented as a categorical value. (Examples: 1, 2)