application_ethics: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data

ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.

Usage

application_ethics

Format

A data frame with 76 rows and 25 variables:

project_id

Unique identifier for each project within the ETH Open Research Data dataset. (Examples: EXPLORE/INSCRIBE, CONTRIBUTE/PYSPM, EXPLORE/COORDINEO)

ethics_human_embryonic_stem_cells

Indication of whether human embryonic stem cells are used in the research project, represented as a categorical value. (Examples: FALSE, NA)

ethics_human_embryos

Indication of whether human embryos are used in the research project, represented as a categorical value. (Examples: FALSE)

ethics_human_participants

Indication of whether human participants are involved in the research project, represented as a categorical value. (Examples: FALSE, TRUE)

ethics_human_interventions

Indication of whether human interventions are involved in the research project, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_clinical_study

Indication of whether the research project involves a clinical study, represented as a categorical value. (Examples: FALSE)

ethics_human_cells_tissues

Indication of whether human cells or tissues are used in the research project, represented as a categorical value. (Examples: FALSE)

ethics_personal_data

Indication of whether personal data is collected or used in the research project, represented as a categorical value. (Examples: FALSE, TRUE)

ethics_personal_data_further_processing

Indication of whether personal data will undergo further processing in the research project, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_personal_data_export_switzerland

Indication of whether personal data is exported from Switzerland in the research project, represented as a categorical value. (Examples: FALSE, NA)

ethics_personal_data_import_switzerland

Indication of whether personal data is imported into Switzerland for the research project, represented as a categorical value. (Examples: FALSE, NA)

ethics_personal_data_criminal

Indication of whether personal data related to criminal records is collected or used in the research project, represented as a categorical value. (Examples: FALSE, NA)

ethics_animals

Indication of whether animals are used in the research project, represented as a categorical value. (Examples: FALSE, TRUE)

ethics_activities_outside_switzerland

Indication of whether research activities are conducted outside of Switzerland, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_outside_countries_ethics_issues

Indication of whether there are ethical issues related to activities conducted in countries outside of Switzerland, represented as a categorical value. (Examples: FALSE, NA)

ethics_local_resources

Indication of whether local resources are utilized in the research project, represented as a categorical value. (Examples: FALSE, NA)

ethics_import_material

Indication of whether materials are imported for the research project, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_export_material

Indication of whether materials are exported from the research project, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_low_middle_income_countries

Indication of whether the research project involves activities in low and middle-income countries, represented as a categorical value. (Examples: FALSE, TRUE)

ethics_individuals_at_risk

Indication of whether the research project involves individuals who are considered at risk, represented as a categorical value. (Examples: FALSE, NA)

ethics_environmental_harm

Indication of whether the research project poses a risk of environmental harm, represented as a categorical value. (Examples: FALSE)

ethics_endangered_species

Indication of whether the research project involves endangered species, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_human_harm

Indication of whether the research project poses a risk of harm to humans, represented as a categorical value. (Examples: FALSE, NA)

ethics_artificial_intelligence

Indication of whether the research project involves ethical considerations related to artificial intelligence, represented as a categorical value. (Examples: FALSE, NA, TRUE)

ethics_other_issues

Indication of whether there are other ethical issues relevant to the research project, represented as a categorical value. (Examples: FALSE, NA, TRUE)