application_metadata: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data

ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.

Usage

application_metadata

Format

A data frame with 76 rows and 90 variables:

document_path

character variable

data_project_category

character variable

data_main_applicant_institution

character variable

data_main_applicant_function

character variable

data_main_applicant_first_name

character variable

data_main_applicant_surname

character variable

data_main_applicant_laboratory_name

character variable

data_main_applicant_postcode

integer variable

data_main_applicant_city

character variable

data_all_applicants_0_eth_domain_institution

character variable

data_all_applicants_0_function_title

character variable

data_all_applicants_0_first_name

character variable

data_all_applicants_0_surname

character variable

data_all_applicants_0_postcode

integer variable

data_all_applicants_1_eth_domain_institution

character variable

data_all_applicants_1_function_title

character variable

data_all_applicants_1_first_name

character variable

data_all_applicants_1_surname

character variable

data_all_applicants_1_postcode

integer variable

data_all_applicants_2_eth_domain_institution

character variable

data_all_applicants_2_function_title

character variable

data_all_applicants_2_first_name

character variable

data_all_applicants_2_surname

character variable

data_all_applicants_2_postcode

integer variable

data_all_applicants_3_eth_domain_institution

character variable

data_all_applicants_3_function_title

character variable

data_all_applicants_3_first_name

character variable

data_all_applicants_3_surname

character variable

data_all_applicants_3_postcode

integer variable

data_project_title

character variable

data_project_acronym

character variable

data_project_abstract

character variable

data_keywords_0

character variable

data_keywords_1

character variable

data_keywords_2

character variable

data_keywords_3

character variable

data_keywords_4

character variable

data_start_date

character variable

data_project_duration_months

integer variable

data_total_budget_requested

numeric variable

data_work_packages_0_wp_identifier

character variable

data_work_packages_0_wp_title

character variable

data_work_packages_1_wp_identifier

character variable

data_work_packages_1_wp_title

character variable

data_work_packages_2_wp_identifier

character variable

data_work_packages_2_wp_title

character variable

data_work_packages_3_wp_identifier

character variable

data_work_packages_3_wp_title

character variable

data_work_packages_4_wp_identifier

character variable

data_work_packages_4_wp_title

character variable

data_all_applicants_4_eth_domain_institution

character variable

data_all_applicants_4_function_title

character variable

data_all_applicants_4_first_name

character variable

data_all_applicants_4_surname

character variable

data_all_applicants_4_postcode

integer variable

data_all_applicants_5_eth_domain_institution

character variable

data_all_applicants_5_function_title

character variable

data_all_applicants_5_first_name

character variable

data_all_applicants_5_surname

character variable

data_all_applicants_5_postcode

integer variable

data_work_packages_5_wp_identifier

character variable

data_work_packages_5_wp_title

character variable

data_work_packages

logical variable

data_all_applicants

logical variable

data_keywords

logical variable

data_work_packages_6_wp_identifier

character variable

data_work_packages_6_wp_title

character variable

data_work_packages_7_wp_identifier

character variable

data_work_packages_7_wp_title

character variable

data_work_packages_8_wp_identifier

character variable

data_work_packages_8_wp_title

character variable

data_work_packages_9_wp_identifier

character variable

data_work_packages_9_wp_title

character variable

data_work_packages_10_wp_identifier

character variable

data_work_packages_10_wp_title

character variable

data_keywords_5

character variable

data_keywords_6

character variable

data_keywords_7

character variable

data_keywords_8

character variable

data_work_packages_11_wp_identifier

character variable

data_work_packages_11_wp_title

character variable

data_work_packages_12_wp_identifier

character variable

data_work_packages_12_wp_title

character variable

data_all_applicants_6_eth_domain_institution

character variable

data_all_applicants_6_function_title

character variable

data_all_applicants_6_first_name

character variable

data_all_applicants_6_surname

character variable

data_all_applicants_6_postcode

integer variable

data_keywords_9

character variable

data_keywords_10

character variable