application_metadata_work_packages: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data

ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.

Usage

application_metadata_work_packages

Format

A data frame with 273 rows and 4 variables:

project_id

Unique identifier for each research project within the ETH Open Research Data dataset. (Examples: CONTRIBUTE/XYT, EXPLORE/OPENNEXUS-E, CONTRIBUTE/FIRE)

index

Sequential numeric identifier for each entry in the dataset. (Range: 0 to 12)

wp_identifier

Identifier for the work package associated with the research project. (Examples: other, WP4, WP3)

wp_title

Title or description of the work package associated with the research project. (Examples: Identification of ORD tools and workflows for publication of FAIR qualitative research data, Data Cleaning; Host two online hackathons and a 2-day workshop focused on data cleaning and capacity building., Organization of a public event at EPFL. Communication)