application_metadata: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data

ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.

Usage

application_metadata

Format

A data frame with 76 rows and 15 variables:

project_id

Unique identifier for each research project within the ETH Open Research Data dataset. (Examples: EXPLORE/NMRPRIME, CONTRIBUTE/CATCHARACTPROVI, CONTRIBUTE/THRACE)

project_category

Classification of the research project based on its purpose or funding type. (Examples: Contribute, Establish, Explore)

main_applicant_institution

Institution of the main applicant associated with the research project. (Examples: Eawag, EPFL, Empa)

main_applicant_function

Position or role of the main applicant within their institution. (Examples: Prof, Professor (Prof. Dr.); Senior Scientist, Senior Researcher / Lecturer)

main_applicant_first_name

Given name of the main applicant associated with the research project. (Examples: Esther, Isaac, Elizabeth)

main_applicant_surname

Family name of the main applicant associated with the research project. (Examples: Rieckermann, Paradis, Holmes-Gentle)

main_applicant_laboratory_name

Name of the laboratory associated with the main applicant of the research project. (Examples: NA, Robotic Systems Lab, Chair of Space Geodesy)

main_applicant_postcode

Postal code of the main applicant's address. (Range: 1015 to 9014)

main_applicant_city

City of residence of the main applicant associated with the research project. (Examples: Dübendorf, Birmensdorf, Villigen PSI)

project_title

Title of the research project describing its focus and objectives. (Examples: BRAINSCOREHUMAN – Making Brain and Behavior Data Accessible for Modelling via Brain-Score, DaTabases and fRont-end wEb-bAsed Software for performance-based natUral hazaRds Engineering, MMS (Masonry MicroStructures database) - A 3D masonry microstructures database for advancing numerical research on irregular stone masonry structures)

project_acronym

Abbreviated identifier for the research project, typically representing its full title. (Examples: THRACE, BOSTORD, OpenSense)

project_abstract

Summary of the research project outlining its objectives, methods, and significance. (Examples: The OPEN-ACTRIS project aims to explore and build up FAIR data chain standards and strategies for atmospheric observations collected in Switzerland as part of the Aerosol, Clouds and Trace Gases Research Infrastructure (ACTRIS). ACTRIS is a pan-European network that aims to deepen our understanding of climate change and air pollution by producing high-quality data on short-lived atmospheric constituents. ACTRIS-Switzerland is the multi-institutional Swiss node of the network. ACTRIS has a comprehensive vision for FAIR data that covers all stages of the research data life cycle through the definition of data levels covering raw measurements, processed data, and elaborated data products. The OPEN-ACTRIS project aims to implement these ORD concepts in the ETH domain and to explore best ORD practice within ACTRIS-Switzerland by combining existing tools and infrastructure from the ETH domain and the ACTRIS community. We will achieve this by building FAIR data chains for the aerosol observations continuously recorded at field measurement stations on the Jungfraujoch and in Payerne, and for the aerosol and gas measurements performed on a campaign-basis in the PSI Atmospheric Chemistry Simulation Chambers., Mobile ground robots have become increasingly popular in academia and various industrial applications. However, unlike other domains like aerial robotics, autonomous driving, and construction, there is currently no high-quality, large-scale dataset or reliable benchmark established in this field, nor the tooling available to do so. Creating such a dataset would be immensely valuable for researchers and developers in fostering research on robust and practical algorithms across diverse environments. Moreover, the development of a standardized benchmarking platform would promote fair comparisons between different approaches, fostering innovation and facilitating the rapid progress of mobile ground robot research. Motivated by this, we propose to collect and share a high-quality, versatile, large-scale robotic dataset, “GrandTour”, with scalable and automated tooling– focusing on legged robots in addition to a set of benchmarks and the necessary tooling., The reusability of particle physics analyses is severely affected by undocumented and manual steps taken by the analysis teams. Starting from typically centrally produced reusable data sets, such high-level physics analyses are largely performed using individually developed analysis scripts and frameworks. While the analysis code is commonly under version control in a git repository, this does not guarantee that it works nor that it yields the results obtained from manual execution. This project aims to provide a prototype solution to these issues for experiments performed at the PSI Laboratory for Particle Physics (CMS, Mu3e, n2EDM) by encoding the missing information directly in the code repository and integrating it with the execution environment to ensure the code’s functionality. This will be achieved by providing templates to automatically test the code and package it in self-contained and portable software container images on the code repository server. These container images will then be distributed on the compute clusters using technology that allows for massively parallel batch processing. Aiming for an excellent user experience to increase adoption, this setup that builds upon the existing infrastructure at PSI will be benchmarked and optimized for a potential future large-scale deployment. Dedicated documentation will be developed, a training course be held at PSI, and the results as well as lessons learned will be presented at an international computing workshop.)

start_date

Date when the research project is scheduled to commence, formatted as YYYY-MM-DD. (Examples: 2023-02-01, 2025-01-06, 2023-10-01)

project_duration_months

Duration of the research project measured in months. (Range: 4 to 36)

total_budget_requested

Total budget amount requested for the research project measured in Swiss Francs. (Range: 15000 to 1500000)