report_metadata: ETH Board Open Research Data (ORD) Program Project Metadata and Report Data

ETH Open Research Data dataset, A colletion of metadata and report data of a CHF 15 million ETH domain program dedicated to increase the visibility and the impact of their research within the scientific community, the economy and society as a whole.

Usage

report_metadata

Format

A data frame with 63 rows and 95 variables:

document_path

character variable

data_project_title

character variable

data_explore_call_number

integer variable

data_project_starting_and_end_date

character variable

data_main_applicant_responsible_for_project

character variable

data_function

character variable

data_laboratory_department

character variable

data_institution_of_main_applicant

character variable

data_report

character variable

data_reporting_period

character variable

data_coapplicants

character variable

data_coapplicants_0_name

character variable

data_coapplicants_0_function

character variable

data_coapplicants_0_institution

character variable

data_coapplicants_0_roles_and_main_tasks

character variable

data_coapplicants_1_name

character variable

data_coapplicants_1_function

character variable

data_coapplicants_1_institution

character variable

data_coapplicants_1_roles_and_main_tasks

character variable

data_coapplicants_2_name

character variable

data_coapplicants_2_function

character variable

data_coapplicants_2_institution

character variable

data_coapplicants_2_roles_and_main_tasks

character variable

data_coapplicants_3_name

character variable

data_coapplicants_3_function

character variable

data_coapplicants_3_institution

character variable

data_coapplicants_3_roles_and_main_tasks

character variable

data_coapplicants_4_name

character variable

data_coapplicants_4_function

character variable

data_coapplicants_4_institution

character variable

data_coapplicants_4_roles_and_main_tasks

character variable

data_coapplicants_5_name

character variable

data_coapplicants_5_function

character variable

data_coapplicants_5_institution

character variable

data_coapplicants_5_roles_and_main_tasks

character variable

data_coapplicants_6_name

character variable

data_coapplicants_6_function

character variable

data_coapplicants_6_institution

character variable

data_coapplicants_6_roles_and_main_tasks

character variable

data_coapplicants_7_name

character variable

data_coapplicants_7_function

character variable

data_coapplicants_7_institution

character variable

data_coapplicants_7_roles_and_main_tasks

character variable

data_coapplicants_8_name

character variable

data_coapplicants_8_function

character variable

data_coapplicants_8_institution

character variable

data_coapplicants_8_roles_and_main_tasks

character variable

data_coapplicants_9_name

character variable

data_coapplicants_9_function

character variable

data_coapplicants_9_institution

character variable

data_coapplicants_9_roles_and_main_tasks

character variable

data_coapplicants_10_name

character variable

data_coapplicants_10_function

logical variable

data_coapplicants_10_institution

character variable

data_coapplicants_10_roles_and_main_tasks

character variable

data_coapplicants_11_name

character variable

data_coapplicants_11_function

logical variable

data_coapplicants_11_institution

character variable

data_coapplicants_11_roles_and_main_tasks

character variable

data_coapplicants_12_name

character variable

data_coapplicants_12_function

logical variable

data_coapplicants_12_institution

character variable

data_coapplicants_12_roles_and_main_tasks

character variable

data_coapplicants_13_name

character variable

data_coapplicants_13_function

logical variable

data_coapplicants_13_institution

character variable

data_coapplicants_13_roles_and_main_tasks

character variable

data_coapplicants_14_name

character variable

data_coapplicants_14_function

logical variable

data_coapplicants_14_institution

logical variable

data_coapplicants_14_roles_and_main_tasks

logical variable

data_coapplicants_15_name

character variable

data_coapplicants_15_function

logical variable

data_coapplicants_15_institution

logical variable

data_coapplicants_15_roles_and_main_tasks

logical variable

data_coapplicants_16_name

character variable

data_coapplicants_16_function

logical variable

data_coapplicants_16_institution

logical variable

data_coapplicants_16_roles_and_main_tasks

logical variable

data_coapplicants_17_name

character variable

data_coapplicants_17_function

logical variable

data_coapplicants_17_institution

logical variable

data_coapplicants_17_roles_and_main_tasks

logical variable

data_coapplicants_18_name

character variable

data_coapplicants_18_function

logical variable

data_coapplicants_18_institution

logical variable

data_coapplicants_18_roles_and_main_tasks

logical variable

data_coapplicants_19_name

character variable

data_coapplicants_19_function

logical variable

data_coapplicants_19_institution

logical variable

data_coapplicants_19_roles_and_main_tasks

logical variable

data_coapplicants_20_name

character variable

data_coapplicants_20_function

logical variable

data_coapplicants_20_institution

logical variable

data_coapplicants_20_roles_and_main_tasks

logical variable